cotabest/pyproject.toml
[tool.poetry]
name = "cotabest"
version = "0.1.0"
description = "technical test"
authors = ["Jose Carlos Jr <zejunior.py@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
flask-marshmallow = "^0.14.0"
Flask-Migrate = "^3.1.0"
marshmallow-sqlalchemy = "^0.28.1"
flask-swagger-ui = "^4.11.1"
flask-swagger = "^0.2.14"

[tool.poetry.dev-dependencies]
Flask = "^2.1.3"
Flask-SQLAlchemy = "^2.5.1"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
